requirements.txt
Django==5.2.8
django-allauth==65.13.1
django-widget-tweaks==1.5.0
django-import-export==4.3.14
django-dbbackup==5.0.1
django-simple-history==3.11.0
django-guardian==3.2.0
django-adminactions==2.4
django-background-tasks==1.2.8
python-dotenv==1.0.1
requests==2.32.4
qrcode==8.2
reportlab==4.4.2
PyPDF2==3.0.1
Pillow==10.4.0

